pyproject.toml
[project]
name = "simple-utils"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.0",
    "easyocr>=1.7.2",
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.12.0.88",
    "passlib[argon2]>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "scikit-image>=0.26.0",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn[standard]>=0.40.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "httpx>=0.28.1",
]

